requirements.txt
Django==1.9.10
django-crispy-forms==1.6.1
django-el-pagination==3.0.1
Pillow==2.9.0
python-social-auth==0.2.21
sorl-thumbnail==12.3